fix(repo-wiki,scc-cache,csn-substitutor): rusqlite spawn_blocking + Arc 共享修复
- repo-wiki/store.rs: rusqlite 调用全部 spawn_blocking 包装(§6.2 红线 2)
  - store.rs +851/- 重构为 async 安全的 SQLite 访问层
  - vector.rs: 内存 KNN 实现(ADR-005 降级,禁用 sqlite-vec)
- scc-cache/wal.rs: WAL 恢复 + spawn_blocking 包装
  - wal.rs +735/- 每个恢复周期使用新 TempDir(§6.2 红线 SQLite WAL 锁竞争)
- csn-substitutor: Arc::clone 修复(§4.4 反模式 5)
  - substitutor.rs: Arc::new(self.chains.clone()) → Arc::clone(&self.chains)
- hcw-window: 稀疏压缩优化(compressor/types/window)
- mlc-engine: L0/L1/L2 记忆层调整
- lsct-tiering: spawn_blocking 包装 + 策略层发布事件
- faae-router: EDSB 概率均衡 + 锁持有修复(B-Crit-1/2/3)
- gsoe-evolution/kvbsr-router/osa-coordinator/parliament: 小幅调整
- chimera-cli: commands/* 与 config/lib 同步
- chimera-tui: +9 行调整
- WHY: §6.2 红线 rusqlite 必须 spawn_blocking + §4.4 Arc 共享修复

+++ crates/parliament/src/ahirt.rs
@@ -434,10 +434,14 @@ impl AhirtRedTeam {
         let mut remediation_suggestions = Vec::new();
 
         // WHY total > 0 守卫:空载荷库时 detection_rate = 0.0,不应误报漏洞
-        // WHY f32→f64 提升:detection_rate(f32) 提升至 f64 与 threshold 比较,无损精度
+        // WHY f32 全程比较:§4.4 #6 红线禁止 f32 隐式转 f64 比较
+        //   (f32→f64 精度膨胀会让 0.4f32 as f64 变为 > 0.4,导致边界误判);
+        //   threshold(f64)→f32 是降精度方向,数值不会膨胀,安全。
+        //   更彻底方案是把 config.detection_rate_threshold 改为 f32,但 RC 阶段
+        //   不允许变更核心 API(§3.1),故采用本降精度转换方案。
         // WHY 配置化:阈值来自 config.detection_rate_threshold,替代硬编码 0.95
         let threshold = self.config.detection_rate_threshold;
-        if stats.total > 0 && (stats.detection_rate as f64) < threshold {
+        if stats.total > 0 && stats.detection_rate < threshold as f32 {
             error!(
                 detection_rate = stats.detection_rate,
                 threshold = threshold,
